[setcover,serial] Add an ROI and document my questions in comments
Also, what type of graph (in Ligra parlance) does SetCover expect as
input? E.g.
* AdjacencyGraph as in /mnt/ceph/benchmarks/swarm/inputs/pbbs/mis/* ?
* WeightedEdgeArray as in /mnt/ceph/benchmarks/swarm/inputs/pbbs/graphs/*

And what has this implementation been tested on?

Side note: we should rename those graph directories according to their
type, not awful names that we have now (this is a vestige of my prior
work).