[setcover] Validation: handle vertices with no incoming edges
This isn't really the spirit of set cover. In a (weirdly directed)
bipartite double cover, a vertex with no incoming edges represents an
element that is not in any set. But that violates the set cover
assumption that the union of all sets is equal to the universe of
elements.

TODO: should we just require symmetric graphs?